update script retrieve_c4ghfile_using_fileid.sh

+++ scripts/retrieve_c4ghfile_using_fileid.sh
@@ -2,13 +2,13 @@
 set -euo pipefail
 
 # This script retrieves c4gh file by concatenation of the header (from sda.files) and the archived file from the archive and then decrypt it 
-s3cmdconf=/data/project-sda/misc/bp-submission/s3cmd-bp-master-archive.conf
-binpath=/data/project-sda/my-sda-tools/scripts/
+s3cmdconf=/data3/project-sda/misc/bp-submission/s3cmd-bp-master-archive.conf
+binpath=/data3/project-sda/my-sda-tools/scripts/
 keyfile=c4gh.sec.pem
 outdir=.
 
 usage="""
-Usage: $0 [OPTIONS] <fileid> 
+Usage: $0 [OPTIONS] <fileid1> [fileid2] ...
 Options:
   -h, --help        Show this help message and exit
   -s3cmdconf <path> Path to the s3cmd configuration file (default: $s3cmdconf)
@@ -22,6 +22,7 @@ if [ "$#" -eq 0 ]; then
     exit 1
 fi
 
+file_ids=()
 while [[ $# -gt 0 ]]; do
     case $1 in
         -h|--help)
@@ -67,8 +68,9 @@ while [[ $# -gt 0 ]]; do
               echo "Error: $outdir is not a directory."
               exit 1
             fi
+            ;;
         *)
-            fileid=$1
+            file_ids+=("$1")
             shift
             ;;
     esac
@@ -88,15 +90,30 @@ if [ ! -f $keyfile ];then
     exit 1
 fi
 
+if ! vault kv get -field=private_key bp-secrets/crypt4gh > /dev/null 2>&1; then
+  echo "Error: Unable to retrieve private key from Vault. Ensure you are logged in and have access."
+  exit 1
+fi
+
+vault kv get -field=private_key bp-secrets/crypt4gh > $keyfile 
+export C4GH_PASSPHRASE=$(vault kv get -field=password bp-secrets/crypt4gh)
 
 tmpdir=$(mktemp -d)
 trap 'rm -rf "$tmpdir"' EXIT
 
 
-bash $binpath/get_header_using_fileid.sh $fileid | xxd -r -p > $tmpdir/$fileid.header.bin
-s3cmd -c $s3cmdconf get s3://archive-2024-01/$fileid  $tmpdir/$fileid.newstorage
-cat $tmpdir/$fileid.header.bin $tmpdir/$fileid.newstorage > $tmpdir/$fileid.c4gh
-crypt4gh decrypt -s $keyfile -f $tmpdir/$fileid.c4gh
+for fileid in "${file_ids[@]}"; do
+    if [[ -z "$fileid" ]]; then
+        echo "Error: Empty file ID provided."
+        exit 1
+    fi
+    echo "Processing file ID: $fileid"
 
-cp $tmpdir/$fileid $outdir/$fileid
-echo "File retrieved and decrypted successfully: $fileid"
+    bash $binpath/get_header_using_fileid.sh $fileid | xxd -r -p > $tmpdir/$fileid.header.bin
+    s3cmd -c $s3cmdconf get s3://archive-2024-01/$fileid  $tmpdir/$fileid.newstorage
+    cat $tmpdir/$fileid.header.bin $tmpdir/$fileid.newstorage > $tmpdir/$fileid.c4gh
+    crypt4gh decrypt -s $keyfile -f $tmpdir/$fileid.c4gh
+
+    cp $tmpdir/$fileid $outdir/$fileid
+    echo "File retrieved and decrypted successfully: $fileid"
+done
